Refactor award views
