--- openprocurement/auctions/core/plugins/awarding_1_0/views.py
@@ -1,682 +1,0 @@
-# -*- coding: utf-8 -*-
-from openprocurement.api.models import (
-    STAND_STILL_TIME,
-    get_now
-)
-from openprocurement.api.utils import (
-    json_view,
-    context_unpack,
-    APIResource,
-    calculate_business_date,
-    get_file,
-    upload_file,
-    update_file_content_type,
-    set_ownership
-)
-from openprocurement.api.validation import (
-    validate_file_update,
-    validate_file_upload,
-    validate_patch_document_data,
-)
-from openprocurement.api.views.complaint_document import STATUS4ROLE
-from openprocurement.auctions.core.utils import (
-    apply_patch,
-    save_auction,
-    add_next_award,
-    check_auction_status
-)
-from openprocurement.auctions.core.validation import (
-    validate_award_data,
-    validate_patch_award_data,
-    validate_patch_complaint_data,
-    validate_complaint_data
-)
-
-
-class AuctionAwardResource(APIResource):
-
-    @json_view(permission='view_auction')
-    def collection_get(self):
-        """Auction Awards List
-
-        Get Awards List
-        ---------------
-
-        Example request to get awards list:
-
-        .. sourcecode:: http
-
-            GET /auctions/4879d3f8ee2443169b5fbbc9f89fa607/awards HTTP/1.1
-            Host: example.com
-            Accept: application/json
-
-        This is what one should expect in response:
-
-        .. sourcecode:: http
-
-            HTTP/1.1 200 OK
-            Content-Type: application/json
-
-            {
-                "data": [
-                    {
-                        "status": "active",
-                        "suppliers": [
-                            {
-                                "id": {
-                                    "name": "Державне управління справами",
-                                    "scheme": "https://ns.openprocurement.org/ua/edrpou",
-                                    "uid": "00037256",
-                                    "uri": "http://www.dus.gov.ua/"
-                                },
-                                "address": {
-                                    "countryName": "Україна",
-                                    "postalCode": "01220",
-                                    "region": "м. Київ",
-                                    "locality": "м. Київ",
-                                    "streetAddress": "вул. Банкова, 11, корпус 1"
-                                }
-                            }
-                        ],
-                        "value": {
-                            "amount": 489,
-                            "currency": "UAH",
-                            "valueAddedTaxIncluded": true
-                        }
-                    }
-                ]
-            }
-
-        """
-        return {'data': [i.serialize("view") for i in self.request.validated['auction'].awards]}
-
-    @json_view(content_type="application/json", permission='create_award', validators=(validate_award_data,))
-    def collection_post(self):
-        """Accept or reject bidder application
-
-        Creating new Award
-        ------------------
-
-        Example request to create award:
-
-        .. sourcecode:: http
-
-            POST /auctions/4879d3f8ee2443169b5fbbc9f89fa607/awards HTTP/1.1
-            Host: example.com
-            Accept: application/json
-
-            {
-                "data": {
-                    "status": "active",
-                    "suppliers": [
-                        {
-                            "id": {
-                                "name": "Державне управління справами",
-                                "scheme": "https://ns.openprocurement.org/ua/edrpou",
-                                "uid": "00037256",
-                                "uri": "http://www.dus.gov.ua/"
-                            },
-                            "address": {
-                                "countryName": "Україна",
-                                "postalCode": "01220",
-                                "region": "м. Київ",
-                                "locality": "м. Київ",
-                                "streetAddress": "вул. Банкова, 11, корпус 1"
-                            }
-                        }
-                    ],
-                    "value": {
-                        "amount": 489,
-                        "currency": "UAH",
-                        "valueAddedTaxIncluded": true
-                    }
-                }
-            }
-
-        This is what one should expect in response:
-
-        .. sourcecode:: http
-
-            HTTP/1.1 201 Created
-            Content-Type: application/json
-
-            {
-                "data": {
-                    "id": "4879d3f8ee2443169b5fbbc9f89fa607",
-                    "date": "2014-10-28T11:44:17.947Z",
-                    "status": "active",
-                    "suppliers": [
-                        {
-                            "id": {
-                                "name": "Державне управління справами",
-                                "scheme": "https://ns.openprocurement.org/ua/edrpou",
-                                "uid": "00037256",
-                                "uri": "http://www.dus.gov.ua/"
-                            },
-                            "address": {
-                                "countryName": "Україна",
-                                "postalCode": "01220",
-                                "region": "м. Київ",
-                                "locality": "м. Київ",
-                                "streetAddress": "вул. Банкова, 11, корпус 1"
-                            }
-                        }
-                    ],
-                    "value": {
-                        "amount": 489,
-                        "currency": "UAH",
-                        "valueAddedTaxIncluded": true
-                    }
-                }
-            }
-
-        """
-        auction = self.request.validated['auction']
-        if auction.status != 'active.qualification':
-            self.request.errors.add('body', 'data', 'Can\'t create award in current ({}) auction status'.format(auction.status))
-            self.request.errors.status = 403
-            return
-        award = self.request.validated['award']
-        if any([i.status != 'active' for i in auction.lots if i.id == award.lotID]):
-            self.request.errors.add('body', 'data', 'Can create award only in active lot status')
-            self.request.errors.status = 403
-            return
-        award.complaintPeriod = {'startDate': get_now().isoformat()}
-        auction.awards.append(award)
-        if save_auction(self.request):
-            self.LOGGER.info('Created auction award {}'.format(award.id),
-                        extra=context_unpack(self.request, {'MESSAGE_ID': 'auction_award_create'}, {'award_id': award.id}))
-            self.request.response.status = 201
-            route = self.request.matched_route.name.replace("collection_", "")
-            self.request.response.headers['Location'] = self.request.current_route_url(_route_name=route, award_id=award.id, _query={})
-            return {'data': award.serialize("view")}
-
-    @json_view(permission='view_auction')
-    def get(self):
-        """Retrieving the award
-
-        Example request for retrieving the award:
-
-        .. sourcecode:: http
-
-            GET /auctions/4879d3f8ee2443169b5fbbc9f89fa607/awards/71b6c23ed8944d688e92a31ec8c3f61a HTTP/1.1
-            Host: example.com
-            Accept: application/json
-
-        And here is the response to be expected:
-
-        .. sourcecode:: http
-
-            HTTP/1.0 200 OK
-            Content-Type: application/json
-
-            {
-                "data": {
-                    "id": "4879d3f8ee2443169b5fbbc9f89fa607",
-                    "date": "2014-10-28T11:44:17.947Z",
-                    "status": "active",
-                    "suppliers": [
-                        {
-                            "id": {
-                                "name": "Державне управління справами",
-                                "scheme": "https://ns.openprocurement.org/ua/edrpou",
-                                "uid": "00037256",
-                                "uri": "http://www.dus.gov.ua/"
-                            },
-                            "address": {
-                                "countryName": "Україна",
-                                "postalCode": "01220",
-                                "region": "м. Київ",
-                                "locality": "м. Київ",
-                                "streetAddress": "вул. Банкова, 11, корпус 1"
-                            }
-                        }
-                    ],
-                    "value": {
-                        "amount": 489,
-                        "currency": "UAH",
-                        "valueAddedTaxIncluded": true
-                    }
-                }
-            }
-
-        """
-        return {'data': self.request.validated['award'].serialize("view")}
-
-    @json_view(content_type="application/json", permission='edit_auction', validators=(validate_patch_award_data,))
-    def patch(self):
-        """Update of award
-
-        Example request to change the award:
-
-        .. sourcecode:: http
-
-            PATCH /auctions/4879d3f8ee2443169b5fbbc9f89fa607/awards/71b6c23ed8944d688e92a31ec8c3f61a HTTP/1.1
-            Host: example.com
-            Accept: application/json
-
-            {
-                "data": {
-                    "value": {
-                        "amount": 600
-                    }
-                }
-            }
-
-        And here is the response to be expected:
-
-        .. sourcecode:: http
-
-            HTTP/1.0 200 OK
-            Content-Type: application/json
-
-            {
-                "data": {
-                    "id": "4879d3f8ee2443169b5fbbc9f89fa607",
-                    "date": "2014-10-28T11:44:17.947Z",
-                    "status": "active",
-                    "suppliers": [
-                        {
-                            "id": {
-                                "name": "Державне управління справами",
-                                "scheme": "https://ns.openprocurement.org/ua/edrpou",
-                                "uid": "00037256",
-                                "uri": "http://www.dus.gov.ua/"
-                            },
-                            "address": {
-                                "countryName": "Україна",
-                                "postalCode": "01220",
-                                "region": "м. Київ",
-                                "locality": "м. Київ",
-                                "streetAddress": "вул. Банкова, 11, корпус 1"
-                            }
-                        }
-                    ],
-                    "value": {
-                        "amount": 600,
-                        "currency": "UAH",
-                        "valueAddedTaxIncluded": true
-                    }
-                }
-            }
-
-        """
-        auction = self.request.validated['auction']
-        if auction.status not in ['active.qualification', 'active.awarded']:
-            self.request.errors.add('body', 'data', 'Can\'t update award in current ({}) auction status'.format(auction.status))
-            self.request.errors.status = 403
-            return
-        award = self.request.context
-        if any([i.status != 'active' for i in auction.lots if i.id == award.lotID]):
-            self.request.errors.add('body', 'data', 'Can update award only in active lot status')
-            self.request.errors.status = 403
-            return
-        award_status = award.status
-        apply_patch(self.request, save=False, src=self.request.context.serialize())
-        if award_status == 'pending' and award.status == 'active':
-            award.complaintPeriod.endDate = calculate_business_date(get_now(), STAND_STILL_TIME, auction, True)
-            auction.contracts.append(type(auction).contracts.model_class({
-                'awardID': award.id,
-                'suppliers': award.suppliers,
-                'value': award.value,
-                'date': get_now(),
-                'items': [i for i in auction.items if i.relatedLot == award.lotID ],
-                'contractID': '{}-{}{}'.format(auction.auctionID, self.server_id, len(auction.contracts) + 1) }))
-            add_next_award(self.request)
-        elif award_status == 'active' and award.status == 'cancelled':
-            now = get_now()
-            if award.complaintPeriod.endDate > now:
-                award.complaintPeriod.endDate = now
-            for j in award.complaints:
-                if j.status not in ['invalid', 'resolved', 'declined']:
-                    j.status = 'cancelled'
-                    j.cancellationReason = 'cancelled'
-                    j.dateCanceled = now
-            for i in auction.contracts:
-                if i.awardID == award.id:
-                    i.status = 'cancelled'
-            add_next_award(self.request)
-        elif award_status == 'pending' and award.status == 'unsuccessful':
-            award.complaintPeriod.endDate = calculate_business_date(get_now(), STAND_STILL_TIME, auction, True)
-            add_next_award(self.request)
-        elif award_status == 'unsuccessful' and award.status == 'cancelled' and any([i.status in ['claim', 'answered', 'pending', 'resolved'] for i in award.complaints]):
-            if auction.status == 'active.awarded':
-                auction.status = 'active.qualification'
-                auction.awardPeriod.endDate = None
-            now = get_now()
-            award.complaintPeriod.endDate = now
-            cancelled_awards = []
-            for i in auction.awards[auction.awards.index(award):]:
-                if i.lotID != award.lotID:
-                    continue
-                i.complaintPeriod.endDate = now
-                i.status = 'cancelled'
-                for j in i.complaints:
-                    if j.status not in ['invalid', 'resolved', 'declined']:
-                        j.status = 'cancelled'
-                        j.cancellationReason = 'cancelled'
-                        j.dateCanceled = now
-                cancelled_awards.append(i.id)
-            for i in auction.contracts:
-                if i.awardID in cancelled_awards:
-                    i.status = 'cancelled'
-            add_next_award(self.request)
-        elif self.request.authenticated_role != 'Administrator' and not(award_status == 'pending' and award.status == 'pending'):
-            self.request.errors.add('body', 'data', 'Can\'t update award in current ({}) status'.format(award_status))
-            self.request.errors.status = 403
-            return
-        if save_auction(self.request):
-            self.LOGGER.info('Updated auction award {}'.format(self.request.context.id),
-                        extra=context_unpack(self.request, {'MESSAGE_ID': 'auction_award_patch'}))
-            return {'data': award.serialize("view")}
-
-
-class AuctionAwardComplaintResource(APIResource):
-
-    @json_view(content_type="application/json", permission='create_award_complaint', validators=(validate_complaint_data,))
-    def collection_post(self):
-        """Post a complaint for award
-        """
-        auction = self.request.validated['auction']
-        if auction.status not in ['active.qualification', 'active.awarded']:
-            self.request.errors.add('body', 'data', 'Can\'t add complaint in current ({}) auction status'.format(auction.status))
-            self.request.errors.status = 403
-            return
-        if any([i.status != 'active' for i in auction.lots if i.id == self.context.lotID]):
-            self.request.errors.add('body', 'data', 'Can add complaint only in active lot status')
-            self.request.errors.status = 403
-            return
-        if self.context.complaintPeriod and \
-           (self.context.complaintPeriod.startDate and self.context.complaintPeriod.startDate > get_now() or
-                self.context.complaintPeriod.endDate and self.context.complaintPeriod.endDate < get_now()):
-            self.request.errors.add('body', 'data', 'Can add complaint only in complaintPeriod')
-            self.request.errors.status = 403
-            return
-        complaint = self.request.validated['complaint']
-        complaint.date = get_now()
-        complaint.relatedLot = self.context.lotID
-        if complaint.status == 'claim':
-            complaint.dateSubmitted = get_now()
-        else:
-            complaint.status = 'draft'
-        complaint.complaintID = '{}.{}{}'.format(auction.auctionID, self.server_id, sum([len(i.complaints) for i in auction.awards], len(auction.complaints)) + 1)
-        set_ownership(complaint, self.request)
-        self.context.complaints.append(complaint)
-        if save_auction(self.request):
-            self.LOGGER.info('Created auction award complaint {}'.format(complaint.id),
-                        extra=context_unpack(self.request, {'MESSAGE_ID': 'auction_award_complaint_create'}, {'complaint_id': complaint.id}))
-            self.request.response.status = 201
-            route = self.request.matched_route.name.replace("collection_", "")
-            self.request.response.headers['Location'] = self.request.current_route_url(_route_name=route, complaint_id=complaint.id, _query={})
-            return {
-                'data': complaint.serialize("view"),
-                'access': {
-                    'token': complaint.owner_token
-                }
-            }
-
-    @json_view(permission='view_auction')
-    def collection_get(self):
-        """List complaints for award
-        """
-        return {'data': [i.serialize("view") for i in self.context.complaints]}
-
-    @json_view(permission='view_auction')
-    def get(self):
-        """Retrieving the complaint for award
-        """
-        return {'data': self.context.serialize("view")}
-
-    @json_view(content_type="application/json", permission='edit_complaint', validators=(validate_patch_complaint_data,))
-    def patch(self):
-        """Post a complaint resolution for award
-        """
-        auction = self.request.validated['auction']
-        if auction.status not in ['active.qualification', 'active.awarded']:
-            self.request.errors.add('body', 'data', 'Can\'t update complaint in current ({}) auction status'.format(auction.status))
-            self.request.errors.status = 403
-            return
-        if any([i.status != 'active' for i in auction.lots if i.id == self.request.validated['award'].lotID]):
-            self.request.errors.add('body', 'data', 'Can update complaint only in active lot status')
-            self.request.errors.status = 403
-            return
-        if self.context.status not in ['draft', 'claim', 'answered', 'pending']:
-            self.request.errors.add('body', 'data', 'Can\'t update complaint in current ({}) status'.format(self.context.status))
-            self.request.errors.status = 403
-            return
-        data = self.request.validated['data']
-        complaintPeriod = self.request.validated['award'].complaintPeriod
-        is_complaintPeriod = complaintPeriod.startDate < get_now() and complaintPeriod.endDate > get_now() if complaintPeriod.endDate else complaintPeriod.startDate < get_now()
-        # complaint_owner
-        if self.request.authenticated_role == 'complaint_owner' and self.context.status in ['draft', 'claim', 'answered', 'pending'] and data.get('status', self.context.status) == 'cancelled':
-            apply_patch(self.request, save=False, src=self.context.serialize())
-            self.context.dateCanceled = get_now()
-        elif self.request.authenticated_role == 'complaint_owner' and is_complaintPeriod and self.context.status == 'draft' and data.get('status', self.context.status) == self.context.status:
-            apply_patch(self.request, save=False, src=self.context.serialize())
-        elif self.request.authenticated_role == 'complaint_owner' and is_complaintPeriod and self.context.status == 'draft' and data.get('status', self.context.status) == 'claim':
-            apply_patch(self.request, save=False, src=self.context.serialize())
-            self.context.dateSubmitted = get_now()
-        elif self.request.authenticated_role == 'complaint_owner' and self.context.status == 'answered' and data.get('status', self.context.status) == self.context.status:
-            apply_patch(self.request, save=False, src=self.context.serialize())
-        elif self.request.authenticated_role == 'complaint_owner' and self.context.status == 'answered' and data.get('satisfied', self.context.satisfied) is True and data.get('status', self.context.status) == 'resolved':
-            apply_patch(self.request, save=False, src=self.context.serialize())
-        elif self.request.authenticated_role == 'complaint_owner' and self.context.status == 'answered' and data.get('satisfied', self.context.satisfied) is False and data.get('status', self.context.status) == 'pending':
-            apply_patch(self.request, save=False, src=self.context.serialize())
-            self.context.type = 'complaint'
-            self.context.dateEscalated = get_now()
-        # auction_owner
-        elif self.request.authenticated_role == 'auction_owner' and self.context.status == 'claim' and data.get('status', self.context.status) == self.context.status:
-            apply_patch(self.request, save=False, src=self.context.serialize())
-        elif self.request.authenticated_role == 'auction_owner' and self.context.status == 'claim' and data.get('resolution', self.context.resolution) and len(data.get('resolution', self.context.resolution or "")) >= 20 and data.get('resolutionType', self.context.resolutionType) and data.get('status', self.context.status) == 'answered':
-            apply_patch(self.request, save=False, src=self.context.serialize())
-            self.context.dateAnswered = get_now()
-        elif self.request.authenticated_role == 'auction_owner' and self.context.status == 'pending':
-            apply_patch(self.request, save=False, src=self.context.serialize())
-        # reviewers
-        elif self.request.authenticated_role == 'reviewers' and self.context.status == 'pending' and data.get('status', self.context.status) == self.context.status:
-            apply_patch(self.request, save=False, src=self.context.serialize())
-        elif self.request.authenticated_role == 'reviewers' and self.context.status == 'pending' and data.get('status', self.context.status) in ['resolved', 'invalid', 'declined']:
-            apply_patch(self.request, save=False, src=self.context.serialize())
-            self.context.dateDecision = get_now()
-        else:
-            self.request.errors.add('body', 'data', 'Can\'t update complaint')
-            self.request.errors.status = 403
-            return
-        if self.context.tendererAction and not self.context.tendererActionDate:
-            self.context.tendererActionDate = get_now()
-        if self.context.status not in ['draft', 'claim', 'answered', 'pending'] and auction.status in ['active.qualification', 'active.awarded']:
-            check_auction_status(self.request)
-        if save_auction(self.request):
-            self.LOGGER.info('Updated auction award complaint {}'.format(self.context.id),
-                        extra=context_unpack(self.request, {'MESSAGE_ID': 'auction_award_complaint_patch'}))
-            return {'data': self.context.serialize("view")}
-
-
-class AuctionAwardComplaintDocumentResource(APIResource):
-
-    @json_view(permission='view_auction')
-    def collection_get(self):
-        """Auction Award Complaint Documents List"""
-        if self.request.params.get('all', ''):
-            collection_data = [i.serialize("view") for i in self.context.documents]
-        else:
-            collection_data = sorted(dict([
-                (i.id, i.serialize("view"))
-                for i in self.context.documents
-            ]).values(), key=lambda i: i['dateModified'])
-        return {'data': collection_data}
-
-    @json_view(permission='edit_complaint', validators=(validate_file_upload,))
-    def collection_post(self):
-        """Auction Award Complaint Document Upload
-        """
-        if self.request.validated['auction_status'] not in ['active.qualification', 'active.awarded']:
-            self.request.errors.add('body', 'data', 'Can\'t add document in current ({}) auction status'.format(self.request.validated['auction_status']))
-            self.request.errors.status = 403
-            return
-        if any([i.status != 'active' for i in self.request.validated['auction'].lots if i.id == self.request.validated['award'].lotID]):
-            self.request.errors.add('body', 'data', 'Can add document only in active lot status')
-            self.request.errors.status = 403
-            return
-        if self.context.status not in STATUS4ROLE.get(self.request.authenticated_role, []):
-            self.request.errors.add('body', 'data', 'Can\'t add document in current ({}) complaint status'.format(self.context.status))
-            self.request.errors.status = 403
-            return
-        document = upload_file(self.request)
-        document.author = self.request.authenticated_role
-        self.context.documents.append(document)
-        if save_auction(self.request):
-            self.LOGGER.info('Created auction award complaint document {}'.format(document.id),
-                        extra=context_unpack(self.request, {'MESSAGE_ID': 'auction_award_complaint_document_create'}, {'document_id': document.id}))
-            self.request.response.status = 201
-            document_route = self.request.matched_route.name.replace("collection_", "")
-            self.request.response.headers['Location'] = self.request.current_route_url(_route_name=document_route, document_id=document.id, _query={})
-            return {'data': document.serialize("view")}
-
-    @json_view(permission='view_auction')
-    def get(self):
-        """Auction Award Complaint Document Read"""
-        if self.request.params.get('download'):
-            return get_file(self.request)
-        document = self.request.validated['document']
-        document_data = document.serialize("view")
-        document_data['previousVersions'] = [
-            i.serialize("view")
-            for i in self.request.validated['documents']
-            if i.url != document.url
-        ]
-        return {'data': document_data}
-
-    @json_view(validators=(validate_file_update,), permission='edit_complaint')
-    def put(self):
-        """Auction Award Complaint Document Update"""
-        if self.request.authenticated_role != self.context.author:
-            self.request.errors.add('url', 'role', 'Can update document only author')
-            self.request.errors.status = 403
-            return
-        if self.request.validated['auction_status'] not in ['active.qualification', 'active.awarded']:
-            self.request.errors.add('body', 'data', 'Can\'t update document in current ({}) auction status'.format(self.request.validated['auction_status']))
-            self.request.errors.status = 403
-            return
-        if any([i.status != 'active' for i in self.request.validated['auction'].lots if i.id == self.request.validated['award'].lotID]):
-            self.request.errors.add('body', 'data', 'Can update document only in active lot status')
-            self.request.errors.status = 403
-            return
-        if self.request.validated['complaint'].status not in STATUS4ROLE.get(self.request.authenticated_role, []):
-            self.request.errors.add('body', 'data', 'Can\'t update document in current ({}) complaint status'.format(self.request.validated['complaint'].status))
-            self.request.errors.status = 403
-            return
-        document = upload_file(self.request)
-        document.author = self.request.authenticated_role
-        self.request.validated['complaint'].documents.append(document)
-        if save_auction(self.request):
-            self.LOGGER.info('Updated auction award complaint document {}'.format(self.request.context.id),
-                        extra=context_unpack(self.request, {'MESSAGE_ID': 'auction_award_complaint_document_put'}))
-            return {'data': document.serialize("view")}
-
-    @json_view(content_type="application/json", validators=(validate_patch_document_data,), permission='edit_complaint')
-    def patch(self):
-        """Auction Award Complaint Document Update"""
-        if self.request.authenticated_role != self.context.author:
-            self.request.errors.add('url', 'role', 'Can update document only author')
-            self.request.errors.status = 403
-            return
-        if self.request.validated['auction_status'] not in ['active.qualification', 'active.awarded']:
-            self.request.errors.add('body', 'data', 'Can\'t update document in current ({}) auction status'.format(self.request.validated['auction_status']))
-            self.request.errors.status = 403
-            return
-        if any([i.status != 'active' for i in self.request.validated['auction'].lots if i.id == self.request.validated['award'].lotID]):
-            self.request.errors.add('body', 'data', 'Can update document only in active lot status')
-            self.request.errors.status = 403
-            return
-        if self.request.validated['complaint'].status not in STATUS4ROLE.get(self.request.authenticated_role, []):
-            self.request.errors.add('body', 'data', 'Can\'t update document in current ({}) complaint status'.format(self.request.validated['complaint'].status))
-            self.request.errors.status = 403
-            return
-        if apply_patch(self.request, src=self.request.context.serialize()):
-            update_file_content_type(self.request)
-            self.LOGGER.info('Updated auction award complaint document {}'.format(self.request.context.id),
-                        extra=context_unpack(self.request, {'MESSAGE_ID': 'auction_award_complaint_document_patch'}))
-            return {'data': self.request.context.serialize("view")}
-
-
-class AuctionAwardDocumentResource(APIResource):
-
-    def validate_award_document(self, operation):
-        if self.request.validated['auction_status'] != 'active.qualification':
-            self.request.errors.add('body', 'data', 'Can\'t {} document in current ({}) auction status'.format(operation,
-                                                                                                              self.request.validated[
-                                                                                                                  'auction_status']))
-            self.request.errors.status = 403
-            return
-        if any([i.status != 'active' for i in self.request.validated['auction'].lots if
-                i.id == self.request.validated['award'].lotID]):
-            self.request.errors.add('body', 'data', 'Can {} document only in active lot status'.format(operation))
-            self.request.errors.status = 403
-            return
-        return True
-
-    @json_view(permission='view_auction')
-    def collection_get(self):
-        """Auction Award Documents List"""
-        if self.request.params.get('all', ''):
-            collection_data = [i.serialize("view") for i in self.context.documents]
-        else:
-            collection_data = sorted(dict([
-                (i.id, i.serialize("view"))
-                for i in self.context.documents
-            ]).values(), key=lambda i: i['dateModified'])
-        return {'data': collection_data}
-
-    @json_view(validators=(validate_file_upload,), permission='edit_auction')
-    def collection_post(self):
-        """Auction Award Document Upload
-        """
-        if not self.validate_award_document('add'):
-            return
-        document = upload_file(self.request)
-        self.context.documents.append(document)
-        if save_auction(self.request):
-            self.LOGGER.info('Created auction award document {}'.format(document.id),
-                        extra=context_unpack(self.request, {'MESSAGE_ID': 'auction_award_document_create'}, {'document_id': document.id}))
-            self.request.response.status = 201
-            document_route = self.request.matched_route.name.replace("collection_", "")
-            self.request.response.headers['Location'] = self.request.current_route_url(_route_name=document_route, document_id=document.id, _query={})
-            return {'data': document.serialize("view")}
-
-    @json_view(permission='view_auction')
-    def get(self):
-        """Auction Award Document Read"""
-        if self.request.params.get('download'):
-            return get_file(self.request)
-        document = self.request.validated['document']
-        document_data = document.serialize("view")
-        document_data['previousVersions'] = [
-            i.serialize("view")
-            for i in self.request.validated['documents']
-            if i.url != document.url
-        ]
-        return {'data': document_data}
-
-    @json_view(validators=(validate_file_update,), permission='edit_auction')
-    def put(self):
-        """Auction Award Document Update"""
-        if not self.validate_award_document('update'):
-            return
-        document = upload_file(self.request)
-        self.request.validated['award'].documents.append(document)
-        if save_auction(self.request):
-            self.LOGGER.info('Updated auction award document {}'.format(self.request.context.id),
-                        extra=context_unpack(self.request, {'MESSAGE_ID': 'auction_award_document_put'}))
-            return {'data': document.serialize("view")}
-
-    @json_view(content_type="application/json", validators=(validate_patch_document_data,), permission='edit_auction')
-    def patch(self):
-        """Auction Award Document Update"""
-        if not self.validate_award_document('update'):
-            return
-        if apply_patch(self.request, src=self.request.context.serialize()):
-            update_file_content_type(self.request)
-            self.LOGGER.info('Updated auction award document {}'.format(self.request.context.id),
-                        extra=context_unpack(self.request, {'MESSAGE_ID': 'auction_award_document_patch'}))
-            return {'data': self.request.context.serialize("view")}
